+++ shared/hedvig-ui/Button/button.tsx
@@ -1,7 +1,6 @@
 import styled from '@emotion/styled'
-import { Spinner } from '@hedvig-ui'
+import { Spacing, Spinner } from '@hedvig-ui'
 import { colorsV3 } from '@hedviginsurance/brand'
-import { Spacing } from 'hedvig-ui/spacing'
 import React from 'react'
 import { lightTheme } from '../themes'
