+++ shared/hedvig-ui/Spacing/spacing.stories.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
+import { Spacing, SpacingSize } from '@hedvig-ui'
 import { colorsV2 } from '@hedviginsurance/brand'
 import { boolean, select } from '@storybook/addon-knobs'
-import { Spacing, SpacingSize } from 'hedvig-ui/spacing'
 import React from 'react'
 
 export default {
