+++ src/components/member/tabs/ClaimsTab.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled'
-import { Button, DateTimePicker, EnumDropdown } from '@hedvig-ui'
+import { Button, DateTimePicker, EnumDropdown, Spacing } from '@hedvig-ui'
 import { ClaimSource, useCreateClaimMutation } from 'api/generated/graphql'
 import { ClaimListHeader } from 'components/claims/claims-list/components/ClaimListHeader'
 import { ClaimListItem } from 'components/claims/claims-list/components/ClaimListItem'
@@ -10,7 +10,6 @@ import {
   LoadingMessage,
   StandaloneMessage,
 } from 'hedvig-ui/animations/standalone-message'
-import { Spacing } from 'hedvig-ui/spacing'
 import { MainHeadline } from 'hedvig-ui/typography'
 import React from 'react'
 import { toast } from 'react-hot-toast'
