+++ src/components/member/tabs/contracts-tab/agreement/ToDate.tsx
@@ -1,11 +1,10 @@
-import { Button, ButtonsGroup, DateTimePicker } from '@hedvig-ui'
+import { Button, ButtonsGroup, DateTimePicker, Spacing } from '@hedvig-ui'
 import { Contract, GenericAgreement } from 'api/generated/graphql'
 import { format } from 'date-fns'
 import {
   changeToDateOptions,
   useChangeToDate,
 } from 'graphql/use-change-to-date-for-agreement'
-import { Spacing } from 'hedvig-ui/spacing'
 import {
   FourthLevelHeadline,
   Paragraph,
