+++ src/components/questions/questions-list/AnswerForm.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled'
-import { Checkbox, Form, FormTextArea, SubmitButton } from '@hedvig-ui'
+import { Checkbox, Form, FormTextArea, Spacing, SubmitButton } from '@hedvig-ui'
 import Grid from '@material-ui/core/Grid'
 import {
   getAnswerQuestionOptions,
@@ -9,7 +9,6 @@ import {
   getMarkQuestionAsResolvedOptions,
   useMarkQuestionAsResolved,
 } from 'graphql/use-mark-question-as-resolved'
-import { Spacing } from 'hedvig-ui/spacing'
 import React from 'react'
 import { FieldValues, FormProvider, useForm } from 'react-hook-form'
 import { toast } from 'react-hot-toast'
