+++ src/features/tools/campaign-codes/components/CampaignCodeFilter.tsx
@@ -1,7 +1,6 @@
-import { InfoContainer, Input } from '@hedvig-ui'
+import { InfoContainer, Input, Spacing } from '@hedvig-ui'
 import { CampaignFilter } from 'api/generated/graphql'
 import { PartnerDropdown } from 'features/tools/campaign-codes/forms/PartnerDropdown'
-import { Spacing } from 'hedvig-ui/spacing'
 import { ThirdLevelHeadline } from 'hedvig-ui/typography'
 import React from 'react'
