+++ src/features/tools/campaign-codes/forms/MonthlyPercentageForm.tsx
@@ -1,4 +1,10 @@
-import { Button, DateTimePicker, Input, SearchableDropdown } from '@hedvig-ui'
+import {
+  Button,
+  DateTimePicker,
+  Input,
+  SearchableDropdown,
+  Spacing,
+} from '@hedvig-ui'
 import { AssignVoucherPercentageDiscount, Scalars } from 'api/generated/graphql'
 import { PartnerDropdown } from 'features/tools/campaign-codes/forms/PartnerDropdown'
 import { getCodeTypeOptions } from 'features/tools/campaign-codes/utils'
@@ -6,7 +12,6 @@ import {
   addPartnerPercentageDiscountCodeOptions,
   useAddPartnerPercentageDiscountCode,
 } from 'graphql/use-add-partner-percentage-discount-code'
-import { Spacing } from 'hedvig-ui/spacing'
 import { Label } from 'hedvig-ui/typography'
 import React from 'react'
 import { toast } from 'react-hot-toast'
